requirements.txt
django==1.11.*
django-hashid-field==1.2.2
djangorestframework==3.6.*
